+++ src/gui.h
@@ -1,0 +1,12 @@
+#ifndef _GUI_H_
+#define _GUI_H_
+
+#include <stdlib.h>
+#include <stdio.h>
+#include <math.h>
+#include "GL/gl.h"
+#include "GL/glut.h"
+#include "boule.h"
+#include "decor.h"
+
+#endif
